Fix language parsing for repeating strings
Having identical strings for different languages on the spreadsheet
used to confuse the language parser, causing it to leak a string from
one language to another. The Flat XML ODF Spreadsheet (.fods) format
used by the languge system can store only one occurrence of each
repeating string of a row, causing the parser to use the next one,
which belongs to another language.

- Fix #44.